stock-advisor/requirements.txt
pathway
yfinance
torch
transformers
accelerate
fastapi
uvicorn
python-dotenv
requests
pandas
numpy